Fix for ordering of some keyword tokens in matching.
Some of the tokens were not placed sensibly - things like 'MODE'
were coloured as 'MOD' and a variable 'E'. This change moves some
around to avoid this.

The file handle related functions (eg PTR#, CLOSE# etc) now include
the `#` as part of the token. Similarly the string variable/functions
include the `$` as part oftheir names now.